refactor: `NodeServer` internal module
